feat: implement admin dashboard statistics and middleware
