+++ resources/views/admin/index.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <!-- Breathing in, I calm body and mind. Breathing out, I smile. - Thich Nhat Hanh -->
+</div>
